C tests: deduplicate some test mains
Deduplicate standard yaep_read_grammar() type tests.
