+++ test/C/common.h
@@ -88,4 +88,33 @@ test_standard_parse (void)
   yaep_free_grammar (g);
 }
 
+static void
+test_standard_read
+  (const char *(*read_terminal) (int *),
+   const char *(*read_rule) (const char ***, const char **, int *, int **))
+{
+  struct grammar *g;
+  struct yaep_tree_node *root;
+  int ambiguous_p;
+
+  if ((g = yaep_create_grammar ()) == NULL)
+    {
+      fprintf (stderr, "yaep_create_grammar: No memory\n");
+      exit (1);
+    }
+  if (yaep_read_grammar (g, 1, read_terminal, read_rule) != 0)
+    {
+      fprintf (stderr, "%s\n", yaep_error_message (g));
+      exit (1);
+    }
+  if (yaep_parse
+      (g, test_read_token, test_syntax_error, test_parse_alloc,
+       test_parse_free, &root, &ambiguous_p))
+    {
+      fprintf (stderr, "yaep parse: %s\n", yaep_error_message (g));
+      exit (1);
+    }
+  yaep_free_grammar (g);
+}
+
 #endif
